Testing test class without widgets
