+++ pubspec.yaml
@@ -59,6 +59,7 @@ dev_dependencies:
   mockito: ^5.4.4
   build_runner: ^2.4.9
   analyzer: ^5.13.0
+  test: ^1.25.7
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
