--- test/mocks/autofill_widget_test.mocks.dart
@@ -1,45 +1,0 @@
-// Mocks generated by Mockito 5.4.4 from annotations
-// in angeleno_project/test/autofill_widget_test.dart.
-// Do not manually edit this file.
-
-// ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'dart:async' as _i3;
-
-import 'package:flutter_keyboard_visibility/src/keyboard_visibility_controller.dart'
-    as _i2;
-import 'package:mockito/mockito.dart' as _i1;
-
-// ignore_for_file: type=lint
-// ignore_for_file: avoid_redundant_argument_values
-// ignore_for_file: avoid_setters_without_getters
-// ignore_for_file: comment_references
-// ignore_for_file: deprecated_member_use
-// ignore_for_file: deprecated_member_use_from_same_package
-// ignore_for_file: implementation_imports
-// ignore_for_file: invalid_use_of_visible_for_testing_member
-// ignore_for_file: prefer_const_constructors
-// ignore_for_file: unnecessary_parenthesis
-// ignore_for_file: camel_case_types
-// ignore_for_file: subtype_of_sealed_class
-
-/// A class which mocks [KeyboardVisibilityController].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockKeyboardVisibilityController extends _i1.Mock
-    implements _i2.KeyboardVisibilityController {
-  MockKeyboardVisibilityController() {
-    _i1.throwOnMissingStub(this);
-  }
-
-  @override
-  _i3.Stream<bool> get onChange => (super.noSuchMethod(
-        Invocation.getter(#onChange),
-        returnValue: _i3.Stream<bool>.empty(),
-      ) as _i3.Stream<bool>);
-
-  @override
-  bool get isVisible => (super.noSuchMethod(
-        Invocation.getter(#isVisible),
-        returnValue: false,
-      ) as bool);
-}
