--- test/mocks/maps_autofill_api_test.mocks.dart
@@ -1,148 +1,0 @@
-// Mocks generated by Mockito 5.4.4 from annotations
-// in angeleno_project/test/maps_autofill_api_test.dart.
-// Do not manually edit this file.
-
-// ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'dart:async' as _i6;
-
-import 'package:angeleno_project/controllers/place_api.dart' as _i4;
-import 'package:angeleno_project/models/autofill_place.dart' as _i3;
-import 'package:angeleno_project/models/autofill_suggestion.dart' as _i7;
-import 'package:http/http.dart' as _i2;
-import 'package:mockito/mockito.dart' as _i1;
-import 'package:mockito/src/dummies.dart' as _i5;
-
-// ignore_for_file: type=lint
-// ignore_for_file: avoid_redundant_argument_values
-// ignore_for_file: avoid_setters_without_getters
-// ignore_for_file: comment_references
-// ignore_for_file: deprecated_member_use
-// ignore_for_file: deprecated_member_use_from_same_package
-// ignore_for_file: implementation_imports
-// ignore_for_file: invalid_use_of_visible_for_testing_member
-// ignore_for_file: prefer_const_constructors
-// ignore_for_file: unnecessary_parenthesis
-// ignore_for_file: camel_case_types
-// ignore_for_file: subtype_of_sealed_class
-
-class _FakeClient_0 extends _i1.SmartFake implements _i2.Client {
-  _FakeClient_0(
-    Object parent,
-    Invocation parentInvocation,
-  ) : super(
-          parent,
-          parentInvocation,
-        );
-}
-
-class _FakeAutofillPlace_1 extends _i1.SmartFake implements _i3.AutofillPlace {
-  _FakeAutofillPlace_1(
-    Object parent,
-    Invocation parentInvocation,
-  ) : super(
-          parent,
-          parentInvocation,
-        );
-}
-
-/// A class which mocks [PlaceAPI].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockPlaceAPI extends _i1.Mock implements _i4.PlaceAPI {
-  MockPlaceAPI() {
-    _i1.throwOnMissingStub(this);
-  }
-
-  @override
-  _i2.Client get client => (super.noSuchMethod(
-        Invocation.getter(#client),
-        returnValue: _FakeClient_0(
-          this,
-          Invocation.getter(#client),
-        ),
-      ) as _i2.Client);
-
-  @override
-  set client(_i2.Client? _client) => super.noSuchMethod(
-        Invocation.setter(
-          #client,
-          _client,
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  String get sessionToken => (super.noSuchMethod(
-        Invocation.getter(#sessionToken),
-        returnValue: _i5.dummyValue<String>(
-          this,
-          Invocation.getter(#sessionToken),
-        ),
-      ) as String);
-
-  @override
-  set sessionToken(String? _sessionToken) => super.noSuchMethod(
-        Invocation.setter(
-          #sessionToken,
-          _sessionToken,
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  int get count => (super.noSuchMethod(
-        Invocation.getter(#count),
-        returnValue: 0,
-      ) as int);
-
-  @override
-  set count(int? _count) => super.noSuchMethod(
-        Invocation.setter(
-          #count,
-          _count,
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  String get apiKey => (super.noSuchMethod(
-        Invocation.getter(#apiKey),
-        returnValue: _i5.dummyValue<String>(
-          this,
-          Invocation.getter(#apiKey),
-        ),
-      ) as String);
-
-  @override
-  _i6.Future<List<_i7.AutofillSuggestion>> fetchSuggestions(
-    String? input,
-    String? lang,
-  ) =>
-      (super.noSuchMethod(
-        Invocation.method(
-          #fetchSuggestions,
-          [
-            input,
-            lang,
-          ],
-        ),
-        returnValue: _i6.Future<List<_i7.AutofillSuggestion>>.value(
-            <_i7.AutofillSuggestion>[]),
-      ) as _i6.Future<List<_i7.AutofillSuggestion>>);
-
-  @override
-  _i6.Future<_i3.AutofillPlace> getPlaceDetailFromId(String? placeId) =>
-      (super.noSuchMethod(
-        Invocation.method(
-          #getPlaceDetailFromId,
-          [placeId],
-        ),
-        returnValue: _i6.Future<_i3.AutofillPlace>.value(_FakeAutofillPlace_1(
-          this,
-          Invocation.method(
-            #getPlaceDetailFromId,
-            [placeId],
-          ),
-        )),
-      ) as _i6.Future<_i3.AutofillPlace>);
-}
