+++ test/mocks/place_api_test.mocks.dart
@@ -1,5 +1,5 @@
 // Mocks generated by Mockito 5.4.4 from annotations
-// in angeleno_project/test/maps_autofill_api_test.dart.
+// in angeleno_project/test/place_api_test.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
